remove BrowserGamesAdapter.kt and move the code to CarouselBannerAdapterDelegate.kt and GameCategoryAdapterDelegate.kt
